server/requirements.txt
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.7.9
cffi==1.17.1
channels==4.3.0
channels_redis==4.3.0
cloudinary==1.44.1
constantly==23.10.4
cryptography==45.0.5
daphne==4.2.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
faiss-cpu==1.12.0
filelock==3.19.1
fsspec==2025.7.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.1
networkx==3.5
numpy==2.3.2
packaging==25.0
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
pyOpenSSL==25.1.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
zope.interface==7.2
requests~=2.32.5